Eliminate (most) "signed and unsigned comparison" warnings from agent code [-Wsign-compare]
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@19018 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/agent/mibgroup/agent/extend.c
@@ -1100,8 +1100,8 @@ _extend_find_entry( netsnmp_request_info       *request,
     oid oid_buf[MAX_OID_LEN];
     int oid_len;
     int i;
-    char *token;
-    int   token_len;
+    char  *token;
+    size_t token_len;
 
     if (!request || !table_info || !table_info->indexes
                  || !table_info->indexes->next_variable) {
